+++ packages/sdks/react-sdk/examples/app/StepUp.tsx
@@ -1,0 +1,47 @@
+import React, { useCallback, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { Descope } from '../../src';
+
+const StepUp = () => {
+  const [errorMessage, setErrorMessage] = useState('');
+
+  const navigate = useNavigate();
+  const onSuccess = useCallback(() => {
+    setTimeout(() => {
+      alert('Step Up Succeed!'); // eslint-disable-line no-alert
+    }, 1);
+    navigate('/');
+  }, [navigate]);
+
+  const onError = useCallback(() => {
+    setErrorMessage('Something went wrong');
+  }, [setErrorMessage]);
+  return (
+	<>
+		<Descope
+        flowId={process.env.DESCOPE_STEP_UP_FLOW_ID}
+        onSuccess={onSuccess}
+        onError={onError}
+        debug={process.env.DESCOPE_DEBUG_MODE === 'true'}
+        theme={process.env.DESCOPE_THEME as any}
+        locale={process.env.DESCOPE_LOCALE as string}
+        redirectUrl={process.env.DESCOPE_REDIRECT_URL}
+        tenant={process.env.DESCOPE_TENANT_ID}
+        telemetryKey={process.env.DESCOPE_TELEMETRY_KEY}
+      />
+		{errorMessage && (
+		<div
+          className="error"
+          style={{
+            margin: 'auto',
+            color: 'red',
+          }}
+        >
+			{errorMessage}
+		</div>
+      )}
+	</>
+  );
+};
+
+export default StepUp;